refractor double to integer

+++ src/main/java/seedu/programmer/logic/commands/EditLabCommand.java
@@ -39,13 +39,13 @@ public class EditLabCommand extends Command {
     public static final String MESSAGE_LAB_NOT_EXISTS = "Lab %d doesn't exist.";
 
     private final int newLabNum;
-    private final Double total;
+    private final Integer total;
     private final Lab original;
 
     /**
      * @param original the lab to be edited.
      * */
-    public EditLabCommand(Lab original, int newLabNum, Double total) {
+    public EditLabCommand(Lab original, int newLabNum, Integer total) {
         this.original = original;
         this.newLabNum = newLabNum;
         this.total = total;
@@ -53,7 +53,7 @@ public class EditLabCommand extends Command {
     /**
      * @param original the lab to be edited.
      * */
-    public EditLabCommand(Lab original, Double total) {
+    public EditLabCommand(Lab original, Integer total) {
         this.original = original;
         this.total = total;
         this.newLabNum = 0;
